retail-agent-system/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai>=2.19.0
openai-agents>=0.10.5
chromadb==0.5.15
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic>=2.12.3
pydantic-settings>=2.5.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
